requirements.txt
aiohttp==3.9.1
aiosignal==1.3.1
annotated-types==0.6.0
anyio==3.7.1
async-timeout==4.0.3
attrs==23.1.0
bidict==0.22.1
certifi==2023.11.17
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
dnspython==2.4.2
exceptiongroup==1.2.0
fastapi==0.105.0
frozenlist==1.4.1
gevent==23.9.1
greenlet==3.0.2
h11==0.14.0
idna==3.6
iniconfig==2.0.0
motor==3.3.2
multidict==6.0.4
packaging==23.2
passlib==1.7.4
pep272-encryption==0.4
pluggy==1.3.0
pycparser==2.21
pycryptodome==3.19.0
pydantic==2.5.2
pydantic_core==2.14.5
PyJWT==2.8.0
pymongo==4.6.1
pytest==7.4.3
python-engineio==4.8.0
python-multipart==0.0.6
python-socketio==5.10.0
requests==2.31.0
simple-websocket==1.0.0
sniffio==1.3.0
starlette==0.27.0
tomli==2.0.1
typing_extensions==4.9.0
urllib3==2.1.0
uvicorn==0.24.0.post1
wsproto==1.2.0
yarl==1.9.4
zope.event==5.0
zope.interface==6.1
